+++ src/mqt/qecc/__init__.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 from ._version import version as __version__
 from .analog_information_decoding.simulators.analog_tannergraph_decoding import AnalogTannergraphDecoder, AtdSimulator
 from .analog_information_decoding.simulators.quasi_single_shot_v2 import QssSimulator
+from .code import CSSCode
 
 from .pyqecc import (
     Code,
@@ -39,4 +40,5 @@ __all__ = [
     "__version__",
     "apply_ecc",
     "sample_iid_pauli_err",
+    "CSSCode",
 ]
