+++ src/mqt/qecc/cc_decoder/color_code.py
@@ -6,7 +6,7 @@ from enum import Enum
 from typing import TYPE_CHECKING
 
 import numpy as np
-from ldpc import mod2
+from ..code import CSSCode
 
 if TYPE_CHECKING:  # pragma: no cover
     import numpy.typing as npt
@@ -19,7 +19,7 @@ class LatticeType(str, Enum):
     SQUARE_OCTAGON = "square_octagon"
 
 
-class ColorCode:
+class ColorCode(CSSCode):
     """A base class for color codes on a three-valent, three-colourable lattice."""
 
     def __init__(self, distance: int, lattice_type: LatticeType) -> None:
@@ -32,10 +32,10 @@ class ColorCode:
         self.lattice_type = lattice_type
         self.add_qubits()
         self.H: npt.NDArray[np.int_] = np.zeros((len(self.ancilla_qubits), len(self.data_qubits)), dtype=int)
+        CSSCode.__init__(self.distance, self.H, self.H)
+        self.L = self.Lz
         self.construct_layout()
-        self.compute_logical()
         self.n = len(self.qubits_to_faces)
-        self.k = self.L.shape[1]
 
     def __hash__(self) -> int:
         """Compute a hash for the color code."""
@@ -53,15 +53,6 @@ class ColorCode:
     def construct_layout(self) -> None:
         """Construct the adjacency lists of the code from the qubits lists. Assumes add_qubits was called."""
 
-    def compute_logical(self) -> None:
-        """Compute the logical matrix L."""
-        ker_hx = mod2.nullspace(self.H)  # compute the kernel basis of hx
-        im_hz_transp = mod2.row_basis(self.H)  # compute the image basis of hz.T
-        log_stack = np.vstack([im_hz_transp, ker_hx])
-        pivots = mod2.row_echelon(log_stack.T)[3]
-        log_op_indices = [i for i in range(im_hz_transp.shape[0], log_stack.shape[0]) if i in pivots]
-        self.L = log_stack[log_op_indices]
-
     def get_syndrome(self, error: npt.NDArray[np.int_]) -> npt.NDArray[np.int_]:
         """Compute the syndrome of the error."""
         return self.H @ error % 2
@@ -69,3 +60,7 @@ class ColorCode:
     def check_if_logical_error(self, residual: npt.NDArray[np.int_]) -> bool:
         """Check if the residual is a logical error."""
         return (self.L @ residual % 2).any() is True
+
+    def is_self_dual(self) -> bool:
+        """Check if the code is self-dual."""
+        return True
