+++ src/mqt/qecc/ft_stateprep/__init__.py
@@ -3,8 +3,6 @@
 from __future__ import annotations
 from .state_prep import StatePrepCircuit, depth_optimal_prep_circuit, gate_optimal_prep_circuit, heuristic_prep_circuit, gate_optimal_verification_circuit, gate_optimal_verification_stabilizers, heuristic_verification_circuit
 from .simulation import NoisyNDFTStatePrepSimulator
-from .code import CSSCode
-
 
 __all__ = [
     'StatePrepCircuit',
@@ -15,5 +13,4 @@ __all__ = [
     'gate_optimal_verification_stabilizers',
     'heuristic_verification_circuit',
     'NoisyNDFTStatePrepSimulator',
-    'CSSCode'    
 ]
